Refactoring, moving audio parser to parsers
